This commit adds a number of changes to allow SnapPy to build with the Microsoft Visual C++ compiler provided in the package "Visual C++ for Python".  The package can be downloaded from: http://www.microsoft.com/en-us/download/details.aspx?id=44266 It is a free release containing components from Visual C++ 2008, which is currently being used to build the python.org distribution.
In addition this commit contains a binary of Togl 2.1 for Windows,
compiled with "Visual C++ for Python".  Currently CyOpenGL checks
the Windows version and it uses the new Togl2.1 in Windows 8 and
8.1.  Earlier versions use the older Togl2.0.  (It ain't broke.)

+++ kernel/kernel_code/chern_simons.c
@@ -744,7 +744,7 @@ static Complex Li2(
 
         s1 = complex_plus(
             s1,
-            complex_real_mult(log(k), w)
+            complex_real_mult(log((double)k), w)
         );
 
         s1 = complex_minus(
